Add warning if output exists

+++ DataExtractor.sh
@@ -3,6 +3,11 @@
 # The purpose of this script is to get the final total energy and dipoles of all .log files
 # We grep
 # This script fails if the .log files have a different structure than expected
+if [ -e Data_Collection.csv ]; then
+        echo "Warning Data_Collection.csv exists and will be overwritten"
+        read -p "Press enter to continue"
+fi
+
 rm Data_Collection.csv
 # Greps all the lines with 'total E' and 'full', delete first every second and then every third line
 grep -E 'total E|full: ' Data1/logs/*.log | sed -n '1~2!p' | sed -n '1~3!p' > Data_Collection.temp
